requirements.txt
Django==2.1.1
psycopg2==2.7.5
djangorestframework==3.8.2
factory-boy==2.11.1
Faker==0.9.0
